refactor: Fixed and reimplemented various screen overlays like for the Aether Portal and Repulsion Shield.
